🎨 style(Store.html): refactor navbar and add panel for store page The navbar has been refactored to improve readability and accessibility. A panel has been added to the store page to provide a better user experience. The panel contains information about the store and its products.
🚀 chore(Store.html): add togglePanel function to script
The togglePanel function is added to the script section of the HTML file. This function toggles the display of the panel element between block and none.

🎨 style(Store.css): add styles for brand value section
This commit adds styles for the brand value section in the Store.css file. The brand value section now has a background color, background size, height, and text color. The panel is now displayed as a block and a line is added to the section with a width of 35rem, margin-bottom of 2%, and a border-bottom of 1rem rgb(250, 246, 246) solid.